feat: a lot of changes, to late to describe them. To be continued

+++ app/staging_methods.py
@@ -1,18 +1,32 @@
 import os
-from shutil import copy
+from shutil import copy, copytree
 
 # Staging method: Copy
-def method_1(local_path, relative_path):
-    copy(local_path, relative_path)
+def local_copy(local_path, relative_path):
+    if os.path.isdir(local_path):
+        # Copy entire directory
+        copytree(local_path, relative_path, dirs_exist_ok=True)
+    elif os.path.isfile(local_path):
+        # Copy single file
+        copy(local_path, relative_path)
+    else:
+        raise ValueError(f"Invalid path: {local_path}")
 
 # Staging method: Symlink
-def method_2(local_path, relative_path):
-    os.symlink(local_path, relative_path)
+def local_symlink(local_path, relative_path):
+    if os.path.isdir(local_path):
+        # Create a symlink for the directory
+        os.symlink(local_path, relative_path, target_is_directory=True)
+    elif os.path.isfile(local_path):
+        # Create a symlink for the file
+        os.symlink(local_path, relative_path)
+    else:
+        raise ValueError(f"Invalid path: {local_path}")
 
 # Add more methods as needed
 
 # Dictionary of available methods
-AVAILABLE_METHODS = {
-    "method_1": method_1,
-    "method_2": method_2,
+AVAILABLE_METHODS: dict[str, callable] = {
+    "local_copy": local_copy,
+    "local_symlink": local_symlink,
 }
